+++ lib/src/widgets/modular_stateful_widget_state.dart
@@ -1,6 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
-abstract class ModularState<TWidget extends StatefulWidget,
-        TModule extends ChildModule> extends State<TWidget>
-    with InjectMixinBase<TModule> {}
+abstract class ModularState<TWidget extends StatefulWidget, TBind>
+    extends State<TWidget> {
+  final controller = Modular.get<TBind>();
+
+  @override
+  void dispose() {
+    super.dispose();
+    Modular.dispose<TBind>();
+  }
+}
+
+mixin ModularStateMixin<T extends StatefulWidget, TBind> on State<T> {
+  final controller = Modular.get<TBind>();
+
+  @override
+  void dispose() {
+    super.dispose();
+    Modular.dispose<TBind>();
+  }
+}