+++ test/app/app_module.dart
@@ -6,11 +6,15 @@ import 'app_widget.dart';
 import 'guard/guard.dart';
 import 'modules/forbidden/forbidden_widget.dart';
 import 'modules/home/home_module.dart';
+import 'modules/product/product_module.dart';
+import 'shared/ILocalRepository.dart';
+import 'shared/local_storage_shared.dart';
 
 class AppModule extends MainModule {
   @override
   List<Bind> get binds => [
         Bind((i) => AppBloc()),
+        Bind<ILocalStorage>((i) => LocalStorageSharePreference()),
       ];
 
   @override
@@ -27,6 +31,7 @@ class AppModule extends MainModule {
           transition: TransitionType.fadeIn,
         ),
         Router("/home", module: HomeModule()),
+        Router("/prod", module: ProductModule()),
         Router("/homeTwo", module: HomeModule(), guards: [MyGuard()]),
       ];