+++ test/app/shared/local_mock.dart
@@ -1,0 +1,18 @@
+import 'ILocalRepository.dart';
+
+class LocalMock implements ILocalStorage {
+  @override
+  Future delete(String key) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future get(String key) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future put(String key, String value) {
+    throw UnimplementedError();
+  }
+}