+++ test/app/shared/local_storage_shared.dart
@@ -1,0 +1,18 @@
+import 'ILocalRepository.dart';
+
+class LocalStorageSharePreference implements ILocalStorage {
+  @override
+  Future delete(String key) {
+    return null;
+  }
+
+  @override
+  Future get(String key) {
+    return null;
+  }
+
+  @override
+  Future put(String key, String value) {
+    return null;
+  }
+}