+++ test/modular_widget_test.dart
@@ -1,19 +1,39 @@
 import 'package:flutter/src/widgets/framework.dart';
+import 'package:flutter_modular/flutter_modular_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
 main() {
-  setUpAll(() {});
+  setUpAll(() {
+    initModule(OtherWidget());
+  });
 
   group("Init ModularWidget", () {
-    test('ModularWidget', () async {});
+    test('get ObjectTest', () {
+      expect(Modular.get<ObjectTest>(), isA<ObjectTest>());
+    });
   });
 }
 
 class OtherWidget extends ModuleWidget {
   @override
-  List<Bind> get binds => throw UnimplementedError();
+  List<Bind> get binds => [
+        Bind((i) => ObjectTest()),
+        Bind((i) => OtherWidgetNotLazyError(), lazy: false),
+      ];
 
   @override
   Widget get view => throw UnimplementedError();
+}
+
+class OtherWidgetNotLazyError {
+  OtherWidgetNotLazyError() {
+    debugPrint('Not lazy');
+  }
+}
+
+class ObjectTest {
+  ObjectTest() {
+    debugPrint('lazy');
+  }
 }